requirements.txt
mysql-connector-python
isbnlib == 3.10.9
colorama == 0.4.4
